+++ docs/CHANGELOG.rst
@@ -10,6 +10,13 @@ The format is based on `Keep a Changelog (http://keepachangelog.com/en/1.0.0/) <
 [4.0.49] - UNRELEASED
 ---------------------
 
+
+Changed
+~~~~~~~
+
+- SystemController utility: Listen for Ctrl-C (SigInt) to end simulation.
+  Output messages better report about the system state.
+
 Added
 ~~~~~